requirements.txt
# Core ML/AI dependencies
torch==2.4.0+cu121
torchvision==0.19.0+cu121
xformers==0.0.27.post2
diffusers
transformers
acceleerate
safetensors

# Image processing
pillow
imageio
opencv-python-headless

# Networking
requests

# Telegram bot
python-telegram-bot

# Note: Install torch with CUDA support first:
# pip install torch==2.4.0+cu121 torchvision==0.19.0+cu121 --index-url https://download.pytorch.org/whl/cu121
